define _addsort/_mulsort to identify its usage in codebase
